+++ .eslintrc.js
@@ -7,6 +7,7 @@ module.exports = {
   },
   plugins: [
     '@typescript-eslint',
+    'prettier',
     'simple-import-sort',
   ],
   extends: [
@@ -16,13 +17,13 @@ module.exports = {
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
     'plugin:import/errors',
     'plugin:import/warnings',
-    'prettier/@typescript-eslint'
+    'plugin:prettier/recommended'
   ],
   rules: {
     'newline-before-return': 'error',
     '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
-    "@typescript-eslint/no-use-before-define": ["error", {"classes": false }],
-    'simple-import-sort/sort': 'error',
+    "@typescript-eslint/no-use-before-define": ["error", { "classes": false }],
+    'simple-import-sort/imports': 'error',
     'sort-imports': 'off',
     'import/first': 'error',
     'import/newline-after-import': 'error',